math: split RbmUtils, so the math namespace imports nothing from biology
RbmUtils carries a biological and a math overload of every BNGL conversion, so
roughly half of a 2191-line class in a biological package was really a math-side
BNGL writer - and the reason MathRuleFactory still imported from biology.

The eight math-typed methods move to cbit.vcell.math.ParticleBnglStringWriter.
The two halves separated cleanly because they were only ever related by name: the
biological overloads take a Model, a Structure or a CompartmentMode, none of
which mean anything in the math layer, and the math ones take none of them.

This is not a P4 violation. The RuleAnalysis contract, which now lives in math, is
itself expressed in terms of BNGL - getReactionBNGLShort(), toBngl(),
getMolecularTypeBNGL() - so the math adapter is required to produce it, and BNGL
is a modelling language rather than one solver's input format.

cbit.vcell.math now imports nothing from the biological or solver layers: 18
violations at the start of this branch, zero now, and MathNamespaceSeparationTest's
accepted list is empty. Its staleness check earned its keep again here, failing
with "These accepted violations no longer exist" until the RbmUtils entry was
removed.

Also records a follow-up in docs/nfsim-abstractions.md. Deleting
RulebasedTransformer.compareOutputs removed the only cross-check of our NFSim XML
against BioNetGen's. It was unreachable, so nothing got worse, but it is worth
resurrecting deliberately as a maintained tool - a tagged test over a small corpus,
or an invocable diagnostic - now that getNFSimXML sits beside NFsimXMLWriter. The
note also observes that neither BNGL half is exercised as an export today, and that
whether the contract should be expressed in BNGL at all is worth asking.

vcell-core Fast group green.
