+++ vcell-core/src/main/java/org/vcell/model/rbm/RbmUtils.java
@@ -1527,15 +1527,6 @@ public class RbmUtils {
 			throw new RuntimeException("Unexpected state for ComponentStatePattern " + componentStatePattern);
 		}
 	}
-	public static String toBnglString(ParticleComponentStatePattern componentStatePattern) {
-		if(componentStatePattern.getParticleComponentStateDefinition() != null) {
-			return "~" + componentStatePattern.getParticleComponentStateDefinition().getName();
-		} else if(componentStatePattern.isAny()) {
-			return "";
-		} else {
-			throw new RuntimeException("Unexpected state for ComponentStatePattern " + componentStatePattern);
-		}
-	}
 	public static String toBnglString(ComponentStateDefinition componentStateDefinition) {
 		if(componentStateDefinition == null) {
 			return "";
@@ -1544,13 +1535,6 @@ public class RbmUtils {
 		}
 	}
 	
-	public static String toBnglString(ParticleComponentStateDefinition componentStateDefinition) {
-		if(componentStateDefinition == null) {
-			return "";
-		} else {
-			return "~" + componentStateDefinition.getName();
-		}
-	}
 	
 	public static String toBnglString(MolecularComponent molecularComponent) {
 		StringBuilder buffer = new StringBuilder(molecularComponent.getName());
@@ -1560,13 +1544,6 @@ public class RbmUtils {
 		return buffer.toString();
 	}
 	
-	public static String toBnglString(ParticleMolecularComponent molecularComponent) {
-		StringBuilder buffer = new StringBuilder(molecularComponent.getName());
-		for (ParticleComponentStateDefinition componentStateDefinition : molecularComponent.getComponentStateDefinitions()) {
-			buffer.append(toBnglString(componentStateDefinition));
-		}
-		return buffer.toString();
-	}
 	
 	public static String toBnglString(MolecularType molecularType, Model model, CompartmentMode compartmentMode) {
 		StringBuilder buffer = new StringBuilder(molecularType.getName());
@@ -1610,20 +1587,6 @@ public class RbmUtils {
 		return buffer.toString();
 	}
 	
-	public static String toBnglString(ParticleMolecularType molecularType) {
-		StringBuilder buffer = new StringBuilder(molecularType.getName());
-		buffer.append("(");
-		List<ParticleMolecularComponent> componentList = molecularType.getComponentList();
-		for (int i = 0; i < componentList.size(); ++ i) {
-			if (i > 0) {
-				buffer.append(",");
-			}
-			buffer.append(toBnglString(componentList.get(i)));
-		}
-		buffer.append(")");
-		return buffer.toString();
-	}
-
 	// ordered by position in the UI - this is what we want to show to the user or to export to file
 	public static String toBnglString(SpeciesPattern speciesPattern, Structure structure, CompartmentMode compartmentMode, int productIndex) {
 		StringBuilder buffer = new StringBuilder();
@@ -1655,21 +1618,6 @@ public class RbmUtils {
 				buffer.append(".");
 			}
 			buffer.append(mtpSortedList.get(i));
-		}
-		return buffer.toString();
-	}
-
-	public static String toBnglString(ParticleSpeciesPattern speciesPattern) {
-		if (speciesPattern == null) {
-			return "";
-		}
-		StringBuilder buffer = new StringBuilder();
-		List<ParticleMolecularTypePattern> molecularTypePatterns = speciesPattern.getParticleMolecularTypePatterns();
-		for (int i = 0; i < molecularTypePatterns.size(); ++ i) {
-			if (i > 0) {
-				buffer.append(".");
-			}
-			buffer.append(toBnglString(molecularTypePatterns.get(i)));
 		}
 		return buffer.toString();
 	}
@@ -1715,28 +1663,6 @@ public class RbmUtils {
 		return buffer.toString();
 	}
 
-	public static String toBnglString(ParticleMolecularTypePattern molecularTypePattern) {
-		StringBuilder buffer = new StringBuilder(molecularTypePattern.getMolecularType().getName());
-		buffer.append("(");
-		List<ParticleMolecularComponentPattern> componentPatterns = molecularTypePattern.getMolecularComponentPatternList();
-		boolean bAddComma = false;
-		for (ParticleMolecularComponentPattern mcp : componentPatterns) {
-//			if (mcp.isImplied()) {
-//				continue;
-//			}
-			if (bAddComma) {
-				buffer.append(",");
-			}
-			buffer.append(toBnglString(mcp));
-			bAddComma = true;
-		}
-		buffer.append(")");
-		if(molecularTypePattern.hasExplicitParticipantMatch()) {
-			buffer.append("%" + molecularTypePattern.getMatchLabel());
-		}
-		return buffer.toString();
-	}
-
 	public static String toBnglString(MolecularComponentPattern molecularComponentPattern) {
 		StringBuilder buffer = new StringBuilder(molecularComponentPattern.getMolecularComponent().getName());
 		if (molecularComponentPattern.getComponentStatePattern() != null) {
@@ -1758,26 +1684,6 @@ public class RbmUtils {
 		return buffer.toString();
 	}
 	
-	public static String toBnglString(ParticleMolecularComponentPattern molecularComponentPattern) {
-		StringBuilder buffer = new StringBuilder(molecularComponentPattern.getMolecularComponent().getName());
-		if (molecularComponentPattern.getComponentStatePattern() != null) {
-			buffer.append(toBnglString(molecularComponentPattern.getComponentStatePattern()));
-		}
-		switch (molecularComponentPattern.getBondType()) {
-		case Exists:
-			buffer.append("!+");
-			break;
-		case None:
-			break;
-		case Possible:
-			buffer.append("!?");
-			break;
-		case Specified:
-			buffer.append("!" + molecularComponentPattern.getBondId());
-			break;
-		}
-		return buffer.toString();
-	}
 	
 	public static String toBnglStringShort(ReactionRule reactionRule, CompartmentMode compartmentMode) {
 		StringBuilder buffer = new StringBuilder();
@@ -1815,29 +1721,6 @@ public class RbmUtils {
 		return buffer.toString();
 	}
 	
-	public static String toBnglStringShort(ParticleJumpProcess particleJumpProcess, List<ParticleSpeciesPattern> reactantSpeciesPatterns, List<ParticleSpeciesPattern> productSpeciesPatterns) {
-		StringBuilder buffer = new StringBuilder();
-
-		for (int i=0; i<reactantSpeciesPatterns.size(); i++){
-			ParticleSpeciesPattern reactantSpeciesPattern = reactantSpeciesPatterns.get(i);
-			if (i>0) {
-				buffer.append(" + ");
-			}
-			buffer.append(toBnglString(reactantSpeciesPattern));
-		}
-		// particleJumpProcesses are not reversible
-		boolean bReversible = false;
-		buffer.append(bReversible ? " <-> " : " -> ");
-		
-		for (int i=0; i<productSpeciesPatterns.size(); i++){
-			ParticleSpeciesPattern productSpeciesPattern = productSpeciesPatterns.get(i);
-			if (i>0) {
-				buffer.append(" + ");
-			}
-			buffer.append(toBnglString(productSpeciesPattern));
-		}
-		return buffer.toString();
-	}
 	
 	public static String toBnglStringLong(ReactionRule reactionRule, CompartmentMode compartmentMode) {
 		String str = reactionRule.getName() + ":\t";
